feat: form mongodb client connect and interaction
DBClient class is defined to handle the Mongo database connection and
interaction with it. An instantiation of the class is exported for use
by other modules. Environment variables are used to populate the host,
port and database name for the database client connection.
isAlive method is used to determine if the client connection is active.
The nbUsers and nbFiles methods give the total number of documents for their
respective 'users' and 'files' collection in the database.