feat: setup api with status and stats endpoints
The express application is setup to only process json requests. It listens
on 5000 or according to the environment variable PORT. It uses the express
Router function to separate the routes; along with AppController.js which
defines the functionality of the routes '/status' and '/stats'.
/status provides feedback on the active connection to the app's redis and
database client. /stats shows the number of users and files stored in the
database.