feat: create new user
A post '/users' route is defined to insert a new user into the database.
The new UsersController.js handles the user endpoint functionality. It
relies on the dbClient which now defines two new methods:
findOneUser - uses mongodb's findOne function and queries the users
collection for a user with the provided filter.
insertOneUser - uses mongodb's insertOne function and inserts a new user
document into the users collection. The field/values provided for the user
are expected to be checked beforehand.