feat: sign-in, sign-out and profile view
Uses basic authentication to authenticate a user. Once authenticated, a
token is generated for the user and all endpoints needing authorization
make use of this token to proceed.

getConnect from AuthController: sign-in the user by validating their
authenticity. An unique token is stored in redis as a key and the user's
id as the value. (the user can thus be authorized with this token
instead of their email and password)
for the user to use the api for 24 hours.
getDisconnect from AuthController: sign-out the user by deleting their
token from redis.
getMe from UsersControllers: retrieves a user's information based on their
token they received from login.