feat: create new files
Define FilesController class to handle the endpoints concerning files.
It has a postUpload static mehod which creates and stores a new file
document to the `files` collection in the mongo database. The
postUploadValidate static method handles the validation and appropriate
http response for the postUpload method.
The findOneFile and insertOneFile methods in the DBClient class are defined
to handle finding a file or inserting a new file into the `files` collection.
The POST /files route is defined for creating a file to be stored.