feat: get one or all files of a user
The `getShow` controller, retrieves one specific file of the authorized user, according to the given file id.
The `getIndex` controller, retrieves all the files of a specific parentId belonging to the authorized user. The results are paginated using mongodb's collection aggregate method. The page and parentId is 0 by default and the pageSize is set to 20.