style: comply with eslint
The functions methods are changed to static functions as they do not make use of personal instance attributes. The classes themself are each exported instead of an instance of theirs. This is done because an instance does not make use of their class's static methods. The functions are invoked with arguments directly for each of the separate routes in the index.js file as they weren't seen as callbacks with the shorter syntax when just the class was exported instead of an instance.