fix: userId stored in database as string
The userId of a file document is now stored as an ObjectId instead of a string. The conversion from string to ObjectId for the userId filed has been made to various function to facilitate their use cases, eg. searching for a file document.
The methods have all been changed to static methods as they do not use any instance attributes.